Filter functionality
Working filter functionality for normal and also <others>

NOTE: Two issues, applying very large filters will change type from notin to in, this is not handled so the filter will be reverse to what we want.
The filter is slightly updated for numbers to filter them correctly and basic functionality works well with dates too, but when too many values are set as a filter it clears the filter. Should be handled by smartly setting the filter to be either in or not in.